org and file changes preparing for dnd work
