+++ src/components/jobcompenents/TJDossierBuild2.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { requestTargetJobDetail } from '../requests/JobRequests';
+import { requestTJDetail } from '../requests/JobRequests';
 import { Link, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import { Droppable, Draggable } from '@hello-pangea/dnd';
