+++ src/components/jobcompenents/TJDossierView.js
@@ -1,4 +1,4 @@
 import { useState, useEffect } from 'react';
-import { requestTargetJobDetail } from '../requests/JobRequests';
+import { requestTJDetail } from '../requests/JobRequests';
 import { Link, useParams } from 'react-router-dom';
 import styled from 'styled-components'
