+++ src/components/jobcompenents/TJSnapshot.js
